+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Updated Otter Export to throw an error if nbconvert<6.0.0 is found
 * Converted Otter Export's PDF via HTML exporter to use nbconvert's WebPDF exporter per [#781](https://github.com/ucbds-infra/otter-grader/issues/781)
 * Removed pdfkit from dependencies
+* Added ability to export PDFs via HTML in grading containers per [#782](https://github.com/ucbds-infra/otter-grader/issues/782)
 
 **v5.6.0:**