+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+**v5.7.0:**
+
+* Switch grading image from Mambaforge to Miniforge due to the [sunestting of Mambaforge](https://conda-forge.org/news/2024/07/29/sunsetting-mambaforge/) per [#846](https://github.com/ucbds-infra/otter-grader/issues/846)
+
 **v5.6.0:**
 
 * Updated Otter Grade to write grading summary for each notebook per [#814](https://github.com/ucbds-infra/otter-grader/issues/814)